chore(db): revert rename
