+++ src/server/database/entities/__tests__/AccountEntity.test.ts
@@ -1,7 +1,7 @@
 import { getManager } from "typeorm";
 import { TestDatabase } from "../../__tests__/helpers";
-import { AccountEntity, createAccount } from "../Account";
-import { createUser } from "../User";
+import { AccountEntity, createAccount } from "../AccountEntity";
+import { createUser } from "../UserEntity";
 
 const testDatabase = new TestDatabase();
