+++ src/server/database/entities/__tests__/UserEntity.test.ts
@@ -1,7 +1,7 @@
 import { getManager } from "typeorm";
 import * as uuid from "uuid";
 import { TestDatabase } from "../../__tests__/helpers";
-import { createUser, UserEntity } from "../User";
+import { createUser, UserEntity } from "../UserEntity";
 
 const testDatabase = new TestDatabase();
