+++ src/server/database/entities/index.ts
@@ -1,9 +1,9 @@
-export * from "./Account";
-export * from "./Session";
-export * from "./User";
+export * from "./AccountEntity";
+export * from "./SessionEntity";
+export * from "./UserEntity";
 
-import { AccountEntity } from "./Account";
-import { SessionEntity } from "./Session";
-import { UserEntity } from "./User";
+import { AccountEntity } from "./AccountEntity";
+import { SessionEntity } from "./SessionEntity";
+import { UserEntity } from "./UserEntity";
 
 export const entities = [AccountEntity, SessionEntity, UserEntity];
